+++ lib/pages/home/bottom_sheets/post_comment_more_actions.dart
@@ -2,6 +2,7 @@ import 'package:Okuna/models/post.dart';
 import 'package:Okuna/models/post_comment.dart';
 import 'package:Okuna/models/user.dart';
 import 'package:Okuna/provider.dart';
+import 'package:Okuna/services/bottom_sheet.dart';
 import 'package:Okuna/services/localization.dart';
 import 'package:Okuna/services/modal_service.dart';
 import 'package:Okuna/services/navigation_service.dart';
@@ -41,6 +42,7 @@ class OBPostCommentMoreActionsBottomSheetState
   NavigationService _navigationService;
   LocalizationService _localizationService;
   ModalService _modalService;
+  BottomSheetService _bottomSheetService;
   bool _requestInProgress;
   CancelableOperation _requestOperation;
 
@@ -64,6 +66,7 @@ class OBPostCommentMoreActionsBottomSheetState
     _navigationService = provider.navigationService;
     _localizationService = provider.localizationService;
     _modalService = provider.modalService;
+    _bottomSheetService = provider.bottomSheetService;
 
     return OBPrimaryColorContainer(
       mainAxisSize: MainAxisSize.min,
@@ -93,7 +96,7 @@ class OBPostCommentMoreActionsBottomSheetState
           title: OBText(
             _localizationService.post__actions_delete_comment,
           ),
-          onTap: _deletePostComment,
+          onTap: _onWantsToDeletePostComment,
         ),
       );
     }
@@ -104,8 +107,9 @@ class OBPostCommentMoreActionsBottomSheetState
           opacity: widget.postComment.isReported ?? false ? 0.5 : 1,
           child: ListTile(
             leading: const OBIcon(OBIcons.report),
-            title: OBText(
-                widget.postComment.isReported ?? false ? _localizationService.post__actions_reported_text : _localizationService.post__actions_report_text),
+            title: OBText(widget.postComment.isReported ?? false
+                ? _localizationService.post__actions_reported_text
+                : _localizationService.post__actions_report_text),
             onTap: _reportPostComment,
           ),
         ),
@@ -139,46 +143,32 @@ class OBPostCommentMoreActionsBottomSheetState
         });
   }
 
-  void _deletePostComment() async {
+  Future _onWantsToDeletePostComment() async {
     if (_requestInProgress) return;
     _setRequestInProgress(true);
     _dismissMoreActions();
-    try {
-      _requestOperation = CancelableOperation.fromFuture(
-          _userService.deletePostComment(
-              postComment: widget.postComment, post: widget.post));
 
-      await _requestOperation.value;
-      if (widget.postComment.parentComment == null)
-        widget.post.decreaseCommentsCount();
-      _toastService.success(message: _localizationService.post__actions_comment_deleted, context: context);
-      if (widget.onPostCommentDeleted != null) {
-        widget.onPostCommentDeleted(widget.postComment);
-      }
-    } catch (error) {
-      _onError(error);
-    } finally {
-      _setRequestInProgress(false);
-    }
+    _bottomSheetService.showConfirmAction(
+        context: context,
+        subtitle: _localizationService.post__actions_delete_comment_description,
+        actionCompleter: (BuildContext context) async {
+          await _userService.deletePostComment(
+              postComment: widget.postComment, post: widget.post);
+          _toastService.success(
+              message: _localizationService.post__actions_comment_deleted,
+              context: context);
+          if (widget.postComment.parentComment == null)
+            widget.post.decreaseCommentsCount();
+          if (widget.onPostCommentDeleted != null) {
+            widget.onPostCommentDeleted(widget.postComment);
+          }
+        });
   }
 
   void _editPostComment() async {
     _dismissMoreActions();
     await _modalService.openExpandedCommenter(
         context: context, post: widget.post, postComment: widget.postComment);
-  }
-
-  void _onError(error) async {
-    if (error is HttpieConnectionRefusedError) {
-      _toastService.error(
-          message: error.toHumanReadableMessage(), context: context);
-    } else if (error is HttpieRequestError) {
-      String errorMessage = await error.toHumanReadableMessage();
-      _toastService.error(message: errorMessage, context: context);
-    } else {
-      _toastService.error(message: _localizationService.error__unknown_error, context: context);
-      throw error;
-    }
   }
 
   void _setRequestInProgress(bool requestInProgress) {
